Added AbstractCopyPayloadRepositoryTests and SQLiteJournalFRCopyPayloadRepositoryTests classes
- Abstracted common test logic into AbstractCopyPayloadRepositoryTests
- Implemented SQLite-specific tests in SQLiteJournalFRCopyPayloadRepositoryTests for acceptance testing
